refactor(build): unify language
rename file to match exports,
build is not strictly an asset concern
