+++ dotcom-rendering/scripts/test/build-check.js
@@ -7,7 +7,7 @@
 
 const find = require('find');
 const loadJsonFile = require('load-json-file');
-const { BUILD_VARIANT } = require('../webpack/bundles');
+const { BUILD_VARIANT } = require('../webpack/builds');
 
 const errorAndThrow = (error) => {
 	console.error(error);
